made the 8ball runner handle filenames, in addition to STDIN

+++ bin/8ball.rb
@@ -11,14 +11,38 @@ class EightBallInterpreter
 
   attr_accessor :files_to_delete
 
+  def initialize
+    @files_to_delete = []
+  end
+
   def extract_filenames
     # return an array of js filenames
     if ARGV.empty?
       code = STDIN.readlines.join("\n")
       [(@files_to_delete = [push_to_temp_js(code)])]
     else
-      ARGV.clone
+      compile_to_js_files(ARGV)
     end
+  end
+
+  def compile_to_js_files(files)
+    js_files = []
+    files.each do |f|
+      if File.exists?(f)
+        dir = File.dirname(File.expand_path(f))
+        js_filename = f.split('/').last.sub(/\.rb$/, ".js")
+        full_filename = "#{dir}/#{js_filename}"
+        js_files.push full_filename
+#        @files_to_delete.push full_filename
+        File.open(full_filename, 'w') do |out|
+          rb = File.read(f)
+          out.write EightBallCompiler.compile_string(rb)
+        end
+      else
+        raise 'file not found'
+      end
+    end
+    js_files
   end
 
   def run
@@ -26,7 +50,7 @@ class EightBallInterpreter
     Open3.popen3("node #{files}") { |stdin, stdout, stderr|
       stdout.each { |line| puts line }
     }
-    FileUtils.rm @files_to_delete
+    FileUtils.rm @files_to_delete if @files_to_delete.size > 0
   end
 
   def push_to_temp_js(code)
